# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: REST API для управления банковскими картами
  contact:
    name: noname-jack
    email: aleksander.moskovchenko@mail.ru
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/cards/block-requests/{id}/reject:
    post:
      tags:
        - Запросы на блокировку карт
      summary: Отклонение запроса на блокировку карты (для администратора)
      description: Отклоняет запрос на блокировку карты по его ID. Доступно только
        администраторам.
      operationId: reject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: withResult
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Запрос успешно отклонён (с возвратом данных)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardBlockRequestDto"
        "204":
          description: Запрос успешно отклонён (без возврата данных)
        "404":
          description: Запрос на блокировку не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
  /api/cards/block-requests/{id}/approve:
    post:
      tags:
        - Запросы на блокировку карт
      summary: Одобрение запроса на блокировку карты (для администратора)
      description: Одобряет запрос на блокировку карты по его ID. Доступно только
        администраторам.
      operationId: approve
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: withResult
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Запрос успешно одобрен (с возвратом данных)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoCardBlockRequestDto"
        "204":
          description: Запрос успешно одобрен (без возврата данных)
        "404":
          description: Запрос на блокировку не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDto"
...
  /api/cards/block-requests:
    get:
      tags:
        - Запросы на блокировку карт
      summary: Получение списка запросов на блокировку карт (для администратора)
      description: Возвращает постраничный список запросов на блокировку карт с возможностью
        фильтрации. Доступно только администраторам.
      operationId: getAllCardBlockRequest
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
        - name: filter
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FilterCardBlockRequest"
      responses:
        "200":
          description: Список запросов успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDtoPagedModelCardBlockRequestDto"
...
components:
  schemas:
...
    ApiResponseDto:
      type: object
      description: Общий формат ответа API
      properties:
        message:
          type: string
          description: Сообщение о результате запроса
          example: success
        data:
          description: "Данные, возвращаемые в ответе"
        error:
          type: string
          description: "Сообщение об ошибке, если запрос неуспешен"
          example: Invalid request
        status:
          type: string
          description: HTTP-статус ответа
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: OK
...
    UserResponse:
      type: object
      description: Данные о пользователе
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: admin
        role:
          type: string
          description: Роль пользователя
          enum:
            - ADMIN
            - USER
          example: ADMIN
...
    ApiResponseDtoCardBlockRequestDto:
      type: object
      description: Общий формат ответа API
      properties:
        message:
          type: string
          description: Сообщение о результате запроса
          example: success
        data:
          $ref: "#/components/schemas/CardBlockRequestDto"
          description: "Данные, возвращаемые в ответе"
        error:
          type: string
          description: "Сообщение об ошибке, если запрос неуспешен"
          example: Invalid request
        status:
          type: string
          description: HTTP-статус ответа
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: OK
    CardBlockRequestDto:
      type: object
      description: Данные о запросе на блокировку карты
      properties:
        id:
          type: integer
          format: int64
          description: ID запроса на блокировку
          example: 1
        card:
          $ref: "#/components/schemas/CardResponse"
          description: Данные о карте
        user:
          $ref: "#/components/schemas/UserResponse"
          description: Данные о пользователе
        status:
          type: string
          description: Статус запроса на блокировку
          enum:
            - PENDING
            - APPROVED
            - REJECTED
          example: PENDING
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания запроса
          example: 2025-09-07T10:59:00
    CardResponse:
      type: object
      description: Данные о карте
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        cardNumber:
          type: string
          description: Номер карты
          example: '****-****-****-3456'
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия карты
          example: 2026-12-31
        status:
          type: string
          description: Статус карты
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: Баланс карты
          example: 1000.5
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    FilterCardBlockRequest:
      type: object
      description: Данные для фильтрации запросов на блокировку карт
      properties:
        userId:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        cardId:
          type: integer
          format: int64
          description: ID карты
          example: 1
        status:
          type: string
          description: Статус запроса на блокировку
          enum:
            - PENDING
            - APPROVED
            - REJECTED
          example: PENDING
    ApiResponseDtoPagedModelCardBlockRequestDto:
      type: object
      description: Общий формат ответа API
      properties:
        message:
          type: string
          description: Сообщение о результате запроса
          example: success
        data:
          $ref: "#/components/schemas/PagedModelCardBlockRequestDto"
          description: "Данные, возвращаемые в ответе"
        error:
          type: string
          description: "Сообщение об ошибке, если запрос неуспешен"
          example: Invalid request
        status:
          type: string
          description: HTTP-статус ответа
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: OK
    PagedModelCardBlockRequestDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardBlockRequestDto"
        page:
          $ref: "#/components/schemas/PageMetadata"
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT токен авторизации
      scheme: bearer
      bearerFormat: JWT